Move StreamingUtil
