+++ clients/client/src/main/java/org/projectnessie/client/ResultStreamPaginator.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.projectnessie.api;
+package org.projectnessie.client;
 
 import java.util.List;
 import java.util.OptionalInt;
