+++ servers/quarkus-server/src/test/java/org/projectnessie/server/TestRest.java
@@ -54,9 +54,9 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.projectnessie.api.ContentsApi;
-import org.projectnessie.api.StreamingUtil;
 import org.projectnessie.api.TreeApi;
 import org.projectnessie.client.NessieClient;
+import org.projectnessie.client.StreamingUtil;
 import org.projectnessie.client.http.HttpClient;
 import org.projectnessie.client.http.HttpClientException;
 import org.projectnessie.client.rest.NessieBadRequestException;
